+++ pairing/circl_bls12381/g2.go
@@ -1,4 +1,4 @@
-package bls12381
+package circl_bls12381
 
 import (
 	"crypto/cipher"
