+++ pairing/circl_bls12381/gt.go
@@ -1,4 +1,4 @@
-package bls12381
+package circl_bls12381
 
 import (
 	"crypto/cipher"
