+++ sign/tbls/tbls_test.go
@@ -3,8 +3,8 @@ package tbls
 import (
 	"testing"
 
-	"github.com/drand/kyber/pairing/bls12381"
 	"github.com/drand/kyber/pairing/bn256"
+	"github.com/drand/kyber/pairing/circl_bls12381"
 	"github.com/drand/kyber/sign/test"
 )
 
@@ -15,7 +15,7 @@ func TestBN256(t *testing.T) {
 }
 
 func TestBLS12381(t *testing.T) {
-	suite := bls12381.NewSuite()
+	suite := circl_bls12381.NewSuite()
 	scheme := NewThresholdSchemeOnG1(suite)
 	test.ThresholdTest(t, suite.G2(), scheme)
 }
