+++ suites/all.go
@@ -3,8 +3,8 @@ package suites
 import (
 	"github.com/drand/kyber/group/edwards25519"
 	"github.com/drand/kyber/group/nist"
-	"github.com/drand/kyber/pairing/bls12381"
 	"github.com/drand/kyber/pairing/bn256"
+	"github.com/drand/kyber/pairing/circl_bls12381"
 )
 
 func init() {
@@ -16,7 +16,7 @@ func init() {
 	register(bn256.NewSuiteG2())
 	register(bn256.NewSuiteGT())
 	register(bn256.NewSuiteBn256())
-	register(bls12381.NewSuiteBLS12381())
+	register(circl_bls12381.NewSuiteBLS12381())
 	// This is a constant time implementation that should be
 	// used as much as possible
 	register(edwards25519.NewBlakeSHA256Ed25519())
